Долг линта, порция 8: горячая перезагрузка в настройках и шаблонах (#858)
Тринадцать вхождений `react-refresh/only-export-components` — восемь файлов
редакторов типов и шаблонов. Храповик 30 → 17.

Уехали в свои модули: сводка типа поля и сборка списка выбираемых типов
(fieldTypeOptions), превью вариантов перечисления, уникальный код при
дублировании, порог версий шаблона, проверка Typst-блоков, реестр редакторов
(typeEditorRegistry — оболочка оставила себе диалог-гард и карточку-секцию),
разбор объявленных параметров шаблона, группировка версий.

Комментарий про тост-указатель уехал вместе с хуком, который его показывает:
в файле панели он описывал бы чужой код.

К порции — живой прогон `types-smoke.mjs`, девять проверок. Он нужен именно
здесь: типы стерегут перенос от опечатки, но не от импорта, уведённого к другому
символу с подходящей сигнатурой — экран остаётся рабочим и показывает не то.
Каждая проверка проверена поломкой того, что стережёт: девять поломок — девять
красных прогонов, и каждый раз краснела нужная.

Не покрыты двое: `uniqueCode` (дублирование типа пишет данные) и
`useMaxTemplateVersions` (порог виден только там, где версий больше порога).
Сказано вслух в e2e/README.md; их перенос сверен чтением и типами.

Проверено: tsc -b, сборка, 583 юнит-теста, храповик 30 → 17 и все восемь живых
прогонов (types 9, fields 15, dialogs 13, settings 8, pages 8, shared-ui 6,
keyboard 11, routing 11).
