+++ src/client/eslint-baseline.json
@@ -1,3 +1,3 @@
 {
-  "react-refresh/only-export-components": 30
+  "react-refresh/only-export-components": 17
 }
