+++ src/client/package.json
@@ -18,7 +18,8 @@
     "test:e2e:dialogs": "node e2e/dialogs-smoke.mjs",
     "test:e2e:shared-ui": "node e2e/shared-ui-smoke.mjs",
     "test:e2e:settings": "node e2e/settings-smoke.mjs",
-    "test:e2e:pages": "node e2e/pages-smoke.mjs"
+    "test:e2e:pages": "node e2e/pages-smoke.mjs",
+    "test:e2e:types": "node e2e/types-smoke.mjs"
   },
   "//engines": "Нижнюю границу Node задаёт react-router 8 (>=22.22.0). Соблюдение обеспечивает engine-strict в .npmrc: ниже порога npm ci падает с EBADENGINE. Образ сборки — node:22-alpine, см. deploy/Dockerfile.web.",
   "engines": {
