+++ src/client/src/features/settings/EnumTypesSection.tsx
@@ -13,14 +13,6 @@ import { nextAutoKey } from './schemaConstants';
 // PrimitiveTypesPage (EnumTypeDetail). Здесь остаются переиспользуемые куски: редактор вариантов
 // и форма создания (в модалке).
 
-/** Человекочитаемое превью вариантов перечисления для строки списка. */
-export function humanEnumPreview(values: EnumOptionDef[]): string {
-  if (values.length === 0) return 'нет вариантов';
-  const labels = values.map(v => v.label).filter(Boolean);
-  const head = labels.slice(0, 3).join(', ');
-  return labels.length > 3 ? `${head} … (+${labels.length - 3})` : head;
-}
-
 // ─── Values editor (список код|имя) ────────────────────────────────────────────
 
 export function ValuesEditor({ values, onChange }: { values: EnumOptionDef[]; onChange: (v: EnumOptionDef[]) => void }) {
