+++ src/client/src/features/settings/SettingsPage.tsx
@@ -14,24 +14,7 @@ import { OrphanBlobsSection } from './OrphanBlobsSection';
 import { MaterialLabelSection } from './MaterialLabelSection';
 import { BackupSection } from './BackupSection';
 import { GithubSettingsSection } from './GithubSettingsSection';
-
-// ─── Settings hook (re-exported for use by other pages) ───────────────────────
-
-export const MAX_VERSIONS_KEY = 'crg.maxTemplateVersions';
-export const DEFAULT_MAX_VERSIONS = 5;
-
-export function useMaxTemplateVersions(): [number, (v: number) => void] {
-  const [value, setValue] = useState(() => {
-    const stored = localStorage.getItem(MAX_VERSIONS_KEY);
-    const parsed = stored ? Number(stored) : NaN;
-    return Number.isFinite(parsed) && parsed >= 2 ? parsed : DEFAULT_MAX_VERSIONS;
-  });
-  function set(v: number) {
-    localStorage.setItem(MAX_VERSIONS_KEY, String(v));
-    setValue(v);
-  }
-  return [value, set];
-}
+import { useMaxTemplateVersions } from './useMaxTemplateVersions';
 
 // ─── Locale settings section ───────────────────────────────────────────────────
